requirements.txt
numpy==1.16.0
tensorflow_gpu==1.14.0
matplotlib==3.1.3
pandas==1.0.1
requests==2.18.4
tqdm==4.43.0
lapsolver==1.1.0
opencv_contrib_python==4.2.0.32
easydict==1.9
Pillow==7.0.0
tensorflow==2.1.0
